experimental/mrpc/examples/python_image_server/pyproject.toml
[build-system]
requires = ["maturin>=1.1,<2.0"]
build-backend = "maturin"

[project]
name = "python_image_server"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]


[tool.maturin]
features = ["pyo3/extension-module"]

[tool.poetry]
name = "python-image-server"
version = "0.1.0"
description = ""
authors = ["daggy1234"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
maturin = "^1.1.0"
pillow = "^10.0.0"

